changed twitterfollowid from int to ulong, I believe that will resolve the userid=0 issue.

+++ Dassanie/Controllers/AlertsController.cs
@@ -79,7 +79,7 @@ namespace Dassanie.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(string FollowerName, int FollowerId, string TriggerWords, bool IncludeLink)
+        public async Task<IActionResult> Create(string FollowerName, ulong FollowerId, string TriggerWords, bool IncludeLink)
         {
             SetupUser();
 
@@ -155,7 +155,7 @@ namespace Dassanie.Controllers
 
 
 
-            alert.TwitterFollowId = (int)followerId;
+            alert.TwitterFollowId = followerId;
             alert.TwitterFollowName = followerName;
             alert.IncludeLink = includeLink;
             alert.TriggerWords = triggerWords;
